one failing test before clone issue
