--- plugins/sentences/src/methods/_participle.js
@@ -1,13 +1,0 @@
-//
-// 'i could drive' -> 'i could have driven'
-
-const useParticiple = function (vb) {
-  if (vb.has('(could|should|would|may)')) {
-    return true
-  }
-  return false
-}
-
-module.exports = {
-  useParticiple,
-}
